pyproject.toml
[project]
name = "vnstock-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "beautifulsoup4==4.13.4",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "et-xmlfile==2.0.0",
    "fake-useragent==2.2.0",
    "fonttools==4.59.0",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.5",
    "numpy==2.3.2",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "psutil==7.0.0",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "requests==2.32.4",
    "seaborn==0.13.2",
    "six==1.17.0",
    "soupsieve==2.7",
    "squarify==0.4.4",
    "tenacity==9.1.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "vnai==2.2.3",
    "vnstock==3.3.0",
    "vnstock-ezchart==0.0.3",
    "wordcloud==1.9.4",
    "zipp==3.23.0",
]
